09Otros/ejemplo_bd/requirements.txt
mysql-connector-python==9.1.0
mysqlclient==2.2.6
